# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/vehicles": {
      "post": {
        "operationId": "VehicleController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterVehicleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Registrar novo veículo",
        "tags": [
          "vehicles"
        ]
      },
      "get": {
        "operationId": "VehicleController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VehicleResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar todos os veículos",
        "tags": [
          "vehicles"
        ]
      }
    },
    "/vehicles/{id}": {
      "get": {
        "operationId": "VehicleController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Buscar veículo por ID",
        "tags": [
          "vehicles"
        ]
      },
      "put": {
        "operationId": "VehicleController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVehicleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar veículo",
        "tags": [
          "vehicles"
        ]
      },
      "delete": {
        "operationId": "VehicleController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remover veículo",
        "tags": [
          "vehicles"
        ]
      }
    },
    "/vehicles/client/{clientId}": {
      "get": {
        "operationId": "VehicleController_findByClient",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VehicleResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Buscar veículos por cliente",
        "tags": [
          "vehicles"
        ]
      }
    },
...
  },
  "info": {
    "title": "Auto Repair Shop OS Management",
    "description": "API para gerenciamento de ordens de serviço de uma oficina mecânica. Sistema completo com gestão de clientes, veículos, peças, orçamentos e ordens de serviço.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RegisterVehicleDto": {
        "type": "object",
        "properties": {
          "plate": {
            "type": "string",
            "example": "ABC-1234"
          },
          "brand": {
            "type": "string",
            "example": "Toyota"
          },
          "model": {
            "type": "string",
            "example": "Corolla"
          },
          "year": {
            "type": "number",
            "example": 2022
          },
          "ownerClientId": {
            "type": "string",
            "example": "uuid-do-cliente"
          }
        },
        "required": [
          "plate",
          "brand",
          "model",
          "year",
          "ownerClientId"
        ]
      },
      "VehicleResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "plate": {
            "type": "string"
          },
          "brand": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "year": {
            "type": "number"
          },
          "ownerClientId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "plate",
          "brand",
          "model",
          "year",
          "ownerClientId"
        ]
      },
      "UpdateVehicleDto": {
        "type": "object",
        "properties": {
          "brand": {
            "type": "string",
            "example": "Honda"
          },
          "model": {
            "type": "string",
            "example": "Civic"
          },
          "year": {
            "type": "number",
            "example": 2023
          }
        }
      },
...
    }
  }
}